services/data-sync-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
APScheduler==3.10.4
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-ulid==3.0.0
